% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_ani.r
\name{anibubble}
\alias{anibubble}
\title{Correlation Animated Bubble plot.}
\usage{
anibubble(
  data,
  x,
  y,
  col.var,
  size.var,
  frame.var,
  fps = 5,
  width = 1900,
  height = 1068.75,
  title = NULL,
  subtitle = NULL,
  xtitle = NULL,
  ytitle = NULL,
  caption = NULL
)
}
\arguments{
\item{data}{input data.frame}

\item{x}{x variable}

\item{y}{y variable}

\item{col.var}{color variable}

\item{size.var}{size variable}

\item{frame.var}{frame variable}

\item{fps}{frame change per second}

\item{width}{width plot size}

\item{height}{height plot size}

\item{title}{main title}

\item{subtitle}{subtitle}

\item{xtitle}{x axis title}

\item{ytitle}{y axis title}

\item{caption}{caption}
}
\value{
plot with ggplot2 grammar
}
\description{
anibubble function will draw  Animated Bubble plot for correlation analysis.
}
\examples{
plot<- anibubble(data=gapminder,x="gdpPercap",y="lifeExp",
                 col.var="continent",size.var = "pop", frame.var = "year")
plot
}
