% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.r
\name{area}
\alias{area}
\title{Deviation Area chart.}
\usage{
area(
  data,
  date,
  value,
  title = NULL,
  subtitle = NULL,
  xtitle = NULL,
  ytitle = NULL,
  caption = NULL
)
}
\arguments{
\item{data}{input dataframe}

\item{date}{date variable}

\item{value}{value variable}

\item{title}{input data.frame}

\item{subtitle}{date variable}

\item{xtitle}{value variable}

\item{ytitle}{input data.frame}

\item{caption}{date variable}
}
\value{
An object of class \code{ggplot}
}
\description{
area function will draw area chart for deviation analysis.
}
\examples{
#prepare data
economics$returns_perc <- c(0, diff(economics$psavert)/economics$psavert[-length(economics$psavert)])
economics<- as.data.frame(economics)

#area chart
plot<- area(data=economics[1:100,],date = "date",value = "returns_perc")
plot

}
