% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areastack.R
\name{areastack}
\alias{areastack}
\title{Trend Stacked Area Chart.}
\usage{
areastack(
  data,
  time,
  value1,
  value2,
  title = NULL,
  subtitle = NULL,
  xtitle = NULL,
  ytitle = NULL,
  caption = NULL
)
}
\arguments{
\item{data}{input data.frame}

\item{time}{time variable}

\item{value1}{value1 variable}

\item{value2}{value2 title}

\item{title}{main title}

\item{subtitle}{subtitle}

\item{xtitle}{x axis title}

\item{ytitle}{y axis title}

\item{caption}{caption}
}
\value{
An object of class \code{ggplot}
}
\description{
areastack function will draw Stacked Area Chart for Trend analysis.
}
\examples{
plot<- areastack(data=as.data.frame(economics),time="date",value1="psavert",value2="uempmed")
plot

}
