% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{boxplot}
\alias{boxplot}
\title{Distribution Box Plot.}
\usage{
boxplot(
  data,
  x,
  y,
  col.var = NULL,
  title = NULL,
  subtitle = NULL,
  xtitle = NULL,
  ytitle = NULL,
  caption = NULL
)
}
\arguments{
\item{data}{input data.frame}

\item{x}{x variable}

\item{col.var}{color variable}

\item{title}{input data.frame}

\item{subtitle}{date variable}

\item{xtitle}{value variable}

\item{ytitle}{input data.frame}

\item{caption}{date variable}
}
\value{
plot with ggplot2 grammar
}
\description{
boxplot function will draw Box Plot for Distribution analysis.
}
\examples{
plot<- boxplot(data=mpg,x="class",col.var="cty")
plot
}
