% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble.r
\name{bubble}
\alias{bubble}
\title{Correlation Bubble plot.}
\usage{
bubble(
  data,
  x,
  y,
  col.var,
  size.var,
  title = NULL,
  subtitle = NULL,
  xtitle = NULL,
  ytitle = NULL,
  caption = NULL
)
}
\arguments{
\item{data}{input data.frame}

\item{x}{x variable}

\item{y}{y variable}

\item{col.var}{color variable}

\item{size.var}{size variable}

\item{title}{main title}

\item{subtitle}{subtitle}

\item{xtitle}{x axis title}

\item{ytitle}{y axis title}

\item{caption}{caption}
}
\value{
plot with ggplot2 grammar
}
\description{
bubble function will draw bubble plot for correlation analysis.
}
\examples{
#prepare data
mpg_select <- mpg[mpg$manufacturer \%in\% c("audi", "ford", "honda", "hyundai"), ]

#Bubble plot
plot<- bubble(data=mpg_select, x="displ", y="cty", col.var="manufacturer", size.var="hwy")
plot
}
