% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{calendar}
\alias{calendar}
\title{Trend Calendar Heat Map.}
\usage{
calendar(data, time, value, title = NULL, subtitle = NULL, caption = NULL)
}
\arguments{
\item{data}{input data.frame}

\item{time}{time variable}

\item{value}{value variable}

\item{title}{main title}

\item{subtitle}{subtitle}

\item{caption}{caption}
}
\value{
plot with ggplot2 grammar
}
\description{
calendar function will draw Calendar Heat Map for Trend analysis.
}
\examples{
#Prepare data
df <- read.csv("https://raw.githubusercontent.com/selva86/datasets/master/yahoo.csv")
df$date <- as.Date(df$date)  # format date
df <- df[df$year >= 2012, ]  # filter years

#calendar
plot<- calendar(data=df,time="date",value="VIX.Close")
plot
}
