% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster}
\alias{cluster}
\title{Groups Clusters}
\usage{
cluster(data, x, y, class, title = NULL, subtitle = NULL, caption = NULL)
}
\arguments{
\item{data}{input data.frame}

\item{x}{x variable}

\item{y}{y variable}

\item{class}{class variable}

\item{title}{main title}

\item{subtitle}{subtitle}

\item{caption}{caption}
}
\value{
plot with ggplot2 grammar
}
\description{
cluster function will draw Clusters for Groups analysis.
}
\examples{
# Compute data with principal components
df <- iris[c(1, 2, 3, 4)]
pca_mod <- prcomp(df)  # compute principal components

# Data frame of principal components
df_pc <- data.frame(pca_mod$x, Species=iris$Species)  # dataframe of principal components
plot<- cluster(data=df_pc,x="PC1",y="PC2",class="Species",
               title="Iris Clustering",
               subtitle="With principal components PC1 and PC2 as X and Y axis",
               caption="Source: Iris")
plot
}
