% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlogram.r
\name{correlogram}
\alias{correlogram}
\title{Correlation Correlogram.}
\usage{
correlogram(data, title = NULL, subtitle = NULL, caption = NULL)
}
\arguments{
\item{data}{input data.frame}

\item{title}{main title}

\item{subtitle}{subtitle}

\item{caption}{caption}
}
\value{
An object of class \code{ggplot}
}
\description{
correlogram function will draw Correlogram for correlation analysis.
}
\examples{
plot<- correlogram(data=mtcars)
plot

}
