% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.r
\name{counts}
\alias{counts}
\title{Correlation Counts chart.}
\usage{
counts(
  data,
  x,
  y,
  title = NULL,
  subtitle = NULL,
  xtitle = NULL,
  ytitle = NULL,
  caption = NULL
)
}
\arguments{
\item{data}{input data.frame}

\item{x}{x variable}

\item{y}{y variable}

\item{title}{main title}

\item{subtitle}{subtitle}

\item{xtitle}{x axis title}

\item{ytitle}{y axis title}

\item{caption}{caption}
}
\value{
plot with ggplot2 grammar
}
\description{
counts function will draw counts chart for correlation analysis.
}
\examples{
plot<- counts(data=mpg, x="cty", y="hwy")
plot
}
