% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrogram.R
\name{dendrogram}
\alias{dendrogram}
\title{Groups Dendrogram.}
\usage{
dendrogram(
  data,
  method = "complete",
  title = NULL,
  subtitle = NULL,
  caption = NULL
)
}
\arguments{
\item{data}{input data.frame}

\item{method}{hierarchical cluster method}

\item{title}{main title}

\item{subtitle}{subtitle}

\item{caption}{caption}

\item{value}{value variable}
}
\value{
plot with ggplot2 grammar
}
\description{
dendrogram function will draw Dendrogram for Groups analysis.
}
\examples{
hc <- hclust(dist(USArrests), "ave")  # hierarchical clustering

dendrogram(data=USArrests)
dendrogram(hc)
}
