% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotbox.R
\name{dotbox}
\alias{dotbox}
\title{Distribution Dot + Box Plot.}
\usage{
dotbox(
  data,
  x,
  y,
  title = NULL,
  subtitle = NULL,
  xtitle = NULL,
  ytitle = NULL,
  caption = NULL
)
}
\arguments{
\item{data}{input data.frame}

\item{x}{x variable}

\item{y}{y variable}

\item{title}{input data.frame}

\item{subtitle}{date variable}

\item{xtitle}{value variable}

\item{ytitle}{input data.frame}

\item{caption}{date variable}
}
\value{
plot with ggplot2 grammar
}
\description{
dotbox function will draw Dot + Box Plot for Distribution analysis.
}
\examples{
plot<- dotbox(data=mpg,x="manufacturer",y="cty")
plot
}
