% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumbbell.R
\name{dumbbell}
\alias{dumbbell}
\title{Ranking Dumbbell Plot.}
\usage{
dumbbell(
  data,
  group,
  dif.var,
  value.var,
  start = "min",
  end = "max",
  title = NULL,
  subtitle = NULL,
  xtitle = NULL,
  ytitle = NULL,
  caption = NULL
)
}
\arguments{
\item{data}{input data.frame}

\item{group}{group variable}

\item{dif.var}{difference variable}

\item{value.var}{value variable}

\item{start}{start value}

\item{end}{end value}

\item{title}{input data.frame}

\item{subtitle}{date variable}

\item{xtitle}{value variable}

\item{ytitle}{input data.frame}

\item{caption}{date variable}
}
\value{
An object of class \code{ggplot}
}
\description{
dumbbell function will draw Dumbbell Plot for Ranking analysis.
}
\examples{
# prep data
gapminder <- gapminder \%>\% group_by(continent,year) \%>\%
             summarise(meangdppercap=mean(gdpPercap)) \%>\% as.data.frame(.)
gapminder[59,"meangdppercap"]<-20088.78

#dumbbell plot
plot<- dumbbell(data=gapminder,group="continent",dif.var="year",value.var="meangdppercap",start=1997,end=2002)
plot

}
