% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enscatter.R
\name{enscatter}
\alias{enscatter}
\title{Correlation scatter plot with encircling.}
\usage{
enscatter(
  data,
  x,
  y,
  col.var = NULL,
  size.var = NULL,
  subset,
  title = NULL,
  subtitle = NULL,
  xtitle = NULL,
  ytitle = NULL,
  caption = NULL
)
}
\arguments{
\item{data}{input data.frame}

\item{x}{x variable}

\item{y}{y variable}

\item{col.var}{color variable}

\item{size.var}{size variable}

\item{subset}{subset list type}

\item{title}{main title}

\item{subtitle}{subtitle}

\item{xtitle}{x axis title}

\item{ytitle}{y axis title}

\item{caption}{caption}
}
\value{
plot with ggplot2 grammar
}
\description{
enscatter function will draw scatter plot with encircling for correlation analysis.
}
\examples{
#make subset
iris_select1<-iris[iris$Species=="setosa",]
iris_select2<-iris[iris$Species=="versicolor",]
iris_select3<-iris[iris$Species=="virginica",]
iris_subset<-list(iris_select1,iris_select2,iris_select3)
#scatter plot with encircling
plot<- enscatter(data=iris, x="Sepal.Length", y="Petal.Length", col.var = "Species", subset = iris_subset)
plot
}
