% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{histogram}
\alias{histogram}
\title{Distribution Histogram.}
\usage{
histogram(
  data,
  x,
  col.var = NULL,
  binwidth = 0.1,
  bins = NULL,
  title = NULL,
  subtitle = NULL,
  xtitle = NULL,
  ytitle = NULL,
  caption = NULL
)
}
\arguments{
\item{data}{input data.frame}

\item{x}{x variable}

\item{col.var}{color variable}

\item{binwidth}{binwidth, default is 0.1}

\item{bins}{number of bins}

\item{title}{input data.frame}

\item{subtitle}{date variable}

\item{xtitle}{value variable}

\item{ytitle}{input data.frame}

\item{caption}{date variable}
}
\value{
plot with ggplot2 grammar
}
\description{
histogram function will draw Histogram for Distribution analysis.
}
\examples{
plot<- histogram(data=mpg,x="displ",col.var="class")
plot
}
