% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jitter.R
\name{jitter}
\alias{jitter}
\title{Correlation Jitter plot.}
\usage{
jitter(
  data,
  x,
  y,
  method = "lm",
  title = NULL,
  subtitle = NULL,
  xtitle = NULL,
  ytitle = NULL,
  caption = NULL
)
}
\arguments{
\item{data}{input data.frame}

\item{x}{x variable}

\item{y}{y variable}

\item{method}{regression method type,default is "loess"}

\item{title}{main title}

\item{subtitle}{subtitle}

\item{xtitle}{x axis title}

\item{ytitle}{y axis title}

\item{caption}{caption}
}
\value{
An object of class \code{ggplot}
}
\description{
jitter function will draw jitter plot for correlation analysis.
}
\examples{
plot<- jitter(data=mpg, x="cty", y="hwy", method="loess")
plot

}
