% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginhist_marginbox.R
\name{margin}
\alias{margin}
\title{Correlation Marginal Histogram/Box plot.}
\usage{
margin(
  data,
  x,
  y,
  type,
  title = NULL,
  subtitle = NULL,
  xtitle = NULL,
  ytitle = NULL,
  caption = NULL
)
}
\arguments{
\item{data}{input data.frame}

\item{x}{x variable}

\item{y}{y variable}

\item{type}{"hist" for histogram,"boxplot" for box plot}

\item{title}{main title}

\item{subtitle}{subtitle}

\item{xtitle}{x axis title}

\item{ytitle}{y axis title}

\item{caption}{caption}
}
\value{
An object of class \code{ggplot}
}
\description{
margin function will draw  Marginal Histogram/Box plot for correlation analysis.
}
\examples{
plot<- margin(data=mpg,x="cty",y="hwy",type="hist")
plot

}
