% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie.R
\name{pie}
\alias{pie}
\title{Composition Pie Chart.}
\usage{
pie(data, title = NULL, caption = NULL)
}
\arguments{
\item{data}{input data.frame}

\item{title}{input data.frame}

\item{caption}{date variable}
}
\value{
An object of class \code{ggplot}
}
\description{
pie function will draw Pie Chart for Composition analysis.
}
\examples{
plot<- pie(data=mpg$class,title = "Title",caption="caption")
plot

}
