% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyramid.R
\name{pyramid}
\alias{pyramid}
\title{Distribution Population Pyramid.}
\usage{
pyramid(
  data,
  x,
  y,
  col.var,
  title = NULL,
  subtitle = NULL,
  xtitle = NULL,
  ytitle = NULL,
  caption = NULL
)
}
\arguments{
\item{data}{input data.frame}

\item{x}{x variable}

\item{y}{y variable}

\item{col.var}{color variable}

\item{title}{input data.frame}

\item{subtitle}{date variable}

\item{xtitle}{value variable}

\item{ytitle}{input data.frame}

\item{caption}{date variable}
}
\value{
plot with ggplot2 grammar
}
\description{
pyramid function will draw Population Pyramid for Distribution analysis.
}
\examples{
# Read data
email_campaign_funnel <- read.csv("https://raw.githubusercontent.com/selva86/datasets/master/email_campaign_funnel.csv")

#Population Pyramid
plot<- pyramid(data=email_campaign_funnel,x="Stage",y="Users",col.var="Gender",
               title = "Title")
plot
}
