% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{scatter}
\alias{scatter}
\title{Correlation scatter plot.}
\usage{
scatter(
  data,
  x,
  y,
  col.var = NULL,
  size.var = NULL,
  method = "loess",
  title = NULL,
  subtitle = NULL,
  xtitle = NULL,
  ytitle = NULL,
  caption = NULL
)
}
\arguments{
\item{data}{input data.frame}

\item{x}{x variable}

\item{y}{y variable}

\item{col.var}{color variable}

\item{size.var}{size variable}

\item{method}{regression method type,default is "loess"}

\item{title}{main title}

\item{subtitle}{subtitle}

\item{xtitle}{x axis title}

\item{ytitle}{y axis title}

\item{caption}{caption}
}
\value{
plot with ggplot2 grammar
}
\description{
scatter function will draw scatter plot for correlation analysis.
}
\examples{
#area chart
plot<- scatter(data=iris, x="Sepal.Length", y="Sepal.Width", col.var = "Species", size.var = "Petal.Length",
               method="lm",
               title="Scatter", subtitle="plot", xtitle="Sepal.Length", ytitle="Sepal.Width", caption="iris")

plot
}
