% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{treemap}
\alias{treemap}
\title{Composition Treemap Chart.}
\usage{
treemap(data, area, fill, label, group, title = NULL, caption = NULL)
}
\arguments{
\item{data}{input data.frame}

\item{area}{area variable}

\item{fill}{fill variable}

\item{label}{label variable}

\item{group}{group variable}

\item{title}{input data.frame}

\item{caption}{date variable}

\item{subtitle}{date variable}
}
\value{
plot with ggplot2 grammar
}
\description{
treemap function will draw Treemap Chart for Composition analysis.
}
\examples{
#prep data
proglangs <- read.csv("https://raw.githubusercontent.com/selva86/datasets/master/proglanguages.csv")

#Treemap
plot<- treemap(data=proglangs, area="value", fill="parent", label="id", group="parent", title="Title", caption="caption")
plot
}
