% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsplot.R
\name{tsplot}
\alias{tsplot}
\title{Trend Time Series Plots.}
\usage{
tsplot(
  data,
  time = NULL,
  value = NULL,
  title = NULL,
  subtitle = NULL,
  xtitle = NULL,
  ytitle = NULL,
  caption = NULL
)
}
\arguments{
\item{data}{input data.frame or ts}

\item{time}{time variable}

\item{value}{value variable}

\item{title}{main title}

\item{subtitle}{subtitle}

\item{xtitle}{x axis title}

\item{ytitle}{y axis title}

\item{caption}{caption}
}
\value{
plot with ggplot2 grammar
}
\description{
tsplot function will draw Time Series Plots for Trend analysis.
}
\examples{
#ts type data
plot<- tsplot(data=AirPassengers)
plot

#data.frame data
plot<- tsplot(data=as.data.frame(economics),time="date",value="uempmed")
plot
}
