% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violin.R
\name{violin}
\alias{violin}
\title{Distribution Violin Plot.}
\usage{
violin(
  data,
  x,
  y,
  title = NULL,
  subtitle = NULL,
  xtitle = NULL,
  ytitle = NULL,
  caption = NULL
)
}
\arguments{
\item{data}{input data.frame}

\item{x}{x variable}

\item{y}{y variable}

\item{title}{input data.frame}

\item{subtitle}{date variable}

\item{xtitle}{value variable}

\item{ytitle}{input data.frame}

\item{caption}{date variable}
}
\value{
plot with ggplot2 grammar
}
\description{
violin function will draw Violin Plot for Distribution analysis.
}
\examples{
plot<- violin(data=mpg,x="class",y="cty")
plot
}
