% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waffle.R
\name{waffle}
\alias{waffle}
\title{Composition Waffle Chart.}
\usage{
waffle(data, title = NULL, subtitle = NULL, caption = NULL)
}
\arguments{
\item{data}{input data.frame}

\item{title}{input data.frame}

\item{subtitle}{date variable}

\item{caption}{date variable}
}
\value{
plot with ggplot2 grammar
}
\description{
waffle function will draw Waffle Chart for Composition analysis.
}
\examples{
plot<- waffle(data=mpg$class,title = "Title",caption="caption")
plot
}
